Refactors and adds new OOPath tests
- Renames XPathTest and related classes and package to OOPathTest
- Splits XPathTest to more classes
- Adds new accumulate tests
- Adds collect test
- Adds fireUntilHalt test in OOPathBindTest class
- Adds some other test cases
