+++ drools-compiler/src/test/java/org/drools/compiler/oopath/BabyBoy.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
@@ -11,29 +11,13 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
- */
+*/
 
-package org.drools.compiler.xpath.graph;
+package org.drools.compiler.oopath;
 
-public class Edge {
+public class BabyBoy extends Child {
 
-    private Vertex<?> inV;
-    private Vertex<?> outV;
-
-    public Vertex<?> getInV() {
-        return inV;
-    }
-
-    public void setInV( Vertex<?> inV ) {
-        this.inV = inV;
-    }
-
-    public Vertex<?> getOutV() {
-        return outV;
-    }
-
-    public void setOutV( Vertex<?> outV ) {
-        this.outV = outV;
+    public BabyBoy(String name, int age) {
+        super(name, age);
     }
 }
