+++ drools-compiler/src/test/java/org/drools/compiler/oopath/Group.java
@@ -1,4 +1,4 @@
-package org.drools.compiler.xpath;
+package org.drools.compiler.oopath;
 
 import java.util.List;
