+++ drools-compiler/src/test/java/org/drools/compiler/oopath/TMDirectory.java
@@ -13,7 +13,7 @@
  * limitations under the License.
 */
 
-package org.drools.compiler.xpath;
+package org.drools.compiler.oopath;
 
 import java.util.List;
