+++ drools-compiler/src/test/java/org/drools/compiler/oopath/graph/Edge.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
@@ -11,37 +11,29 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
-*/
+ *
+ */
 
-package org.drools.compiler.xpath;
+package org.drools.compiler.oopath.graph;
 
-import org.drools.core.phreak.ReactiveList;
+public class Edge {
 
-import java.util.List;
+    private Vertex<?> inV;
+    private Vertex<?> outV;
 
-public class Child extends Person {
-
-    private String mother;
-
-    private final List<Toy> toys = new ReactiveList<Toy>();
-
-    public Child(String name, int age) {
-        super(name, age);
+    public Vertex<?> getInV() {
+        return inV;
     }
 
-    public List<Toy> getToys() {
-        return toys;
+    public void setInV( Vertex<?> inV ) {
+        this.inV = inV;
     }
 
-    public void addToy(Toy toy) {
-        toys.add(toy);
+    public Vertex<?> getOutV() {
+        return outV;
     }
 
-    public String getMother() {
-        return mother;
-    }
-
-    public void setMother(String mother) {
-        this.mother = mother;
+    public void setOutV( Vertex<?> outV ) {
+        this.outV = outV;
     }
 }
